NFC Move some logic from pyodide.ts to api.ts
I think this is a bit tidier.
