+++ src/js/types.ts
@@ -238,8 +238,13 @@ export interface Module {
   noWasmDecoding: boolean;
   quit: (status: number, toThrow: Error) => void;
   preRun: { (): void }[];
-  print: (a: string) => void;
-  printErr: (a: string) => void;
+  print?: (a: string) => void;
+  printErr?: (a: string) => void;
+  arguments: string[];
+  API: API;
+  postRun: ((a: Module) => void) | ((a: Module) => void)[];
+  locateFile: (file: string) => string;
+  exited?: {toThrow: any};
   ENV: { [key: string]: string };
   PATH: any;
   TTY: any;
@@ -328,4 +333,10 @@ export interface API {
 
   makePublicAPI: () => PyodideInterface;
   _Comlink: any;
+
+  dsodir: string;
+  sys: PyProxy;
+  os: PyProxy;
+
+  finalizeBootstrap: () => void;
 }
